.nlogo files

+++ src/example/scala/netlogo/slider/SliderNetlogoConnected.scala
@@ -13,7 +13,7 @@ import org.nlogo.core.CompilerException
  */
 
 /** contains slider name and slider default value */
-case class Default(val a: Int, val name: String)
+case class Default(val a: Float, val name: String)
 
 class mySliderConnectedAgent(sliderModel: SliderModel) extends NetlogoSliderAgent(sliderModel)with Simple {
   override def sliderHandle(value: Double, min: Double, increment: Double, max: Double, buttonReleased: Boolean) = {
